google text shown but didn't sent blob to server
