+++ index.html
@@ -12,11 +12,10 @@
 		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
 		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
 
-		<link href="styles/main.css" rel="stylesheet" />
+		<link href="assets/styles/main.css" rel="stylesheet" />
 		<script src="http://code.jquery.com/jquery-1.11.1.min.js" type="text/javascript"></script>
-
 		
-
+		<script src="assets/js/WavAudioEncoder.min.js" type="text/javascript"></script>
 	</head>
 
 	<body>
@@ -69,7 +68,7 @@
 			<a id="audioDownload"></a>
 		</p>
 
-		<script src="audioRecording.js" type="text/javascript"></script>
+		<script src="assets/js/audioRecording.js" type="text/javascript"></script>
 
 		<script type="text/javascript">
             // Connect a websocket to the server. The port notation is post-
