--- styles/main.css
@@ -1,10 +1,0 @@
-body {
-	font-family: "helvetica neue";
-	color: #666;
-	padding: 0 10%;
-}
-
-h1 {
-	font-size: 20px;
-	margin-bottom: 15px;
-}
